refactor: TIO-97: rename file to be consistent
